+++ ui/src/components/rules/SingleDistributionForm.tsx
