+++ ui/src/components/rules/SortableRule.tsx
@@ -1,20 +1,21 @@
 import { useSortable } from '@dnd-kit/sortable';
-import { IEvaluatable } from '~/types/Evaluatable';
 import { IFlag } from '~/types/Flag';
 import { ISegment } from '~/types/Segment';
 import Rule from './Rule';
 import { IRule } from '~/types/Rule';
+import { IVariant } from '~/types/Variant';
 
 type SortableRuleProps = {
   flag: IFlag;
   rule: IRule;
   segments: ISegment[];
+  variants: IVariant[];
   onSuccess: () => void;
   onDelete: () => void;
 };
 
 export default function SortableRule(props: SortableRuleProps) {
-  const { flag, rule, segments, onSuccess, onDelete } = props;
+  const { flag, rule, segments, variants, onSuccess, onDelete } = props;
   const {
     isDragging,
     attributes,
@@ -46,6 +47,7 @@ export default function SortableRule(props: SortableRuleProps) {
       flag={flag}
       rule={rule}
       segments={segments}
+      variants={variants}
       onSuccess={onSuccess}
       onDelete={onDelete}
     />
